+++ packages/carousel/src/Carousel/Carousel.stories.tsx
@@ -9,7 +9,7 @@ export default {
   title: 'Carousel/Carousel',
 } as Meta;
 
-const Template = ({ totalSlides, visibleSlides, step, showArrows }) => (
+const Template = ({ totalSlides, visibleSlides, showArrows }) => (
   <>
     <TextPairing variant="textpairing-header-2XL" mb="spacer-8">
       <TextPairing.Label>Rapid Solutions with Blocks</TextPairing.Label>
@@ -23,7 +23,6 @@ const Template = ({ totalSlides, visibleSlides, step, showArrows }) => (
       naturalSlideWidth={1}
       naturalSlideHeight={1}
       isIntrinsicHeight
-      step={step}
       showArrows={showArrows}
     >
       <Carousel.Slider>
@@ -33,7 +32,7 @@ const Template = ({ totalSlides, visibleSlides, step, showArrows }) => (
             <Carousel.Slide index={index} key={index}>
               <Card outlined>
                 <TextPairing>
-                  <TextPairing.Label>Testimonial {index}</TextPairing.Label>
+                  <TextPairing.Label>Testimonial {index + 1}</TextPairing.Label>
                   <TextPairing.SubLabel>
                     Working with this company has been a game-changer for our business. Their solutions are fast and
                     effective, and they always exceed our expectations.
@@ -48,7 +47,7 @@ const Template = ({ totalSlides, visibleSlides, step, showArrows }) => (
 );
 
 export const Playground = Template.bind({});
-Playground.args = { visibleSlides: 2, totalSlides: 6, step: 2 };
+Playground.args = { visibleSlides: 2, totalSlides: 6 };
 
 export const WithArrows = Template.bind({});
-WithArrows.args = { visibleSlides: 2, totalSlides: 6, step: 2, showArrows: true };
+WithArrows.args = { visibleSlides: 2, totalSlides: 6, showArrows: true };