+++ packages/carousel/src/Carousel/Carousel.test.tsx
@@ -25,4 +25,71 @@ describe('Carousel', () => {
     );
     expect(screen.queryByText('foo')).not.toBeInTheDocument();
   });
+
+  test('should render Carousel with dots', () => {
+    renderWithProviders(
+      <Carousel visibleSlides={1} naturalSlideWidth={1} naturalSlideHeight={1} showDots>
+        <Carousel.Slider>
+          <Carousel.Slide index={0}>
+            <Text>Label</Text>
+          </Carousel.Slide>
+        </Carousel.Slider>
+      </Carousel>
+    );
+
+    expect(screen.queryAllByRole('button', { name: 'slide dot' }).length).toBe(1);
+  });
+
+  test('should render Carousel with arrows', () => {
+    renderWithProviders(
+      <Carousel visibleSlides={1} naturalSlideWidth={1} naturalSlideHeight={1} showArrows>
+        <Carousel.Slider>
+          <Carousel.Slide index={0}>
+            <Text>Label</Text>
+          </Carousel.Slide>
+          <Carousel.Slide index={1}>
+            <Text>Label 1</Text>
+          </Carousel.Slide>
+        </Carousel.Slider>
+      </Carousel>
+    );
+
+    expect(screen.getByLabelText('back')).toBeInTheDocument();
+    expect(screen.getByLabelText('next')).toBeInTheDocument();
+  });
+
+  test('should render Carousel with arrows and dots', () => {
+    renderWithProviders(
+      <Carousel visibleSlides={1} naturalSlideWidth={1} naturalSlideHeight={1} showArrows showDots>
+        <Carousel.Slider>
+          <Carousel.Slide index={0}>
+            <Text>Label</Text>
+          </Carousel.Slide>
+          <Carousel.Slide index={1}>
+            <Text>Label</Text>
+          </Carousel.Slide>
+        </Carousel.Slider>
+      </Carousel>
+    );
+
+    expect(screen.queryAllByRole('button', { name: 'slide dot' }).length).toBe(2);
+    expect(screen.getByLabelText('back')).toBeInTheDocument();
+    expect(screen.getByLabelText('next')).toBeInTheDocument();
+  });
+
+  test('should group Carousel dots', async () => {
+    renderWithProviders(
+      <Carousel visibleSlides={4} naturalSlideWidth={1} naturalSlideHeight={1} showDots>
+        <Carousel.Slider data-testid="carousel-slider">
+          {new Array(13).fill(null).map((_, index) => (
+            <Carousel.Slide key={index} index={index}>
+              <Text>Slide {index + 1}</Text>
+            </Carousel.Slide>
+          ))}
+        </Carousel.Slider>
+      </Carousel>
+    );
+
+    expect(screen.queryAllByRole('button', { name: 'slide dot' }).length).toBe(4);
+  });
 });