+++ packages/cml/src/replaceCmlPlaceholders/replaceCmlPlaceholders.ts
@@ -77,7 +77,17 @@ const replacePropValues = (mdx: string, values: Record<string, string> = {}): st
       case 'boolean':
       case 'number':
         searchValue = match;
-        newValue = `{${propValue}}`;
+        newValue = `{${propValue ?? defaultValue}}`;
+        break;
+      case 'json':
+        searchValue = match;
+
+        if (propValue && typeof propValue === 'object') {
+          newValue = `{${JSON.stringify(propValue)}}`;
+        } else {
+          newValue = `{${propValue}}`;
+        }
+
         break;
       case 'list':
         searchValue = match;