+++ packages/components/src/components/navigation/Stepper/types.ts
@@ -1,0 +1,31 @@
+import { FC } from 'react';
+import {
+  StepDescription,
+  StepIcon,
+  StepIndicator,
+  StepNumber,
+  StepProps,
+  StepSeparator,
+  StepStatus,
+  StepTitle,
+  StepperProps as ChakraStepperProps,
+} from '@chakra-ui/react';
+
+export interface StepperProps extends Omit<ChakraStepperProps, 'size' | 'variant'> {
+  size?: never;
+  variant?: never;
+}
+
+export interface StepStaticMembers {
+  Icon: typeof StepIcon;
+  Indicator: typeof StepIndicator;
+  Number: typeof StepNumber;
+  Separator: typeof StepSeparator;
+  Status: typeof StepStatus;
+  Title: typeof StepTitle;
+  Description: typeof StepDescription;
+}
+
+export interface StepperStaticMembers {
+  Step: FC<StepProps> & StepStaticMembers;
+}