+++ packages/signalwire/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Change Log
 
+## 0.3.28
+
+## 0.3.27
+
+## 0.3.26
+
+## 0.3.25
+
+## 0.3.24
+
+### Patch Changes
+
+- 6ee90cd: fixes Dropdown types and IconButton forwardRef
+
+## 0.3.23
+
 ## 0.3.22
 
 ## 0.3.21